IPTV_v5.4_MAX_AGGRESSION/vps/ape-realtime-guardian/requirements.txt
psutil>=5.9.0
pyyaml>=6.0
aiofiles>=23.0
prometheus_client>=0.20.0
